Split test_newmodels into multiple files
