+++ tests/unit/test_reta_month_offense_subcat_summary.py
@@ -1,178 +1,5 @@
-# -*- coding: utf-8 -*-
+from crime_data.common.newmodels import RetaMonthOffenseSubcatSummary
 
-from crime_data.common.newmodels import (RetaMonthOffenseSubcatSummary,
-                                         AgencyParticipation,
-                                         ParticipationRate,
-                                         CdeAgency)
-import pytest
-
-## Check these tests
-class TestAgencyParticipation:
-    def test_for_agency_not_reporting(self, app):
-        q = AgencyParticipation.query
-        q = q.filter(AgencyParticipation.year == 2014)
-        q = q.filter(AgencyParticipation.agency_id == 17380).one()
-        assert q.reported == 0
-        assert q.months_reported == 0
-        assert q.nibrs_reported == 0
-        assert q.nibrs_months_reported == 0
-
-    def test_for_agency_covered_by_another(self, app):
-        q = AgencyParticipation.query
-        q = q.filter(AgencyParticipation.year == 2014)
-        q = q.filter(AgencyParticipation.agency_id == 17391).one()
-        assert q.reported == 0
-        assert q.months_reported == 0
-        assert q.nibrs_reported == 0
-        assert q.nibrs_months_reported == 0
-        assert q.covered == 1
-        assert q.participated == 1
-        assert q.nibrs_participated == 1
-
-    def test_for_agency_in_nibrs_month(self, app):
-        q = AgencyParticipation.query
-        q = q.filter(AgencyParticipation.year == 2014)
-        q = q.filter(AgencyParticipation.agency_id == 17381).one()
-        assert q.reported == 1
-        assert q.months_reported == 12
-        assert q.nibrs_reported == 1
-        assert q.nibrs_months_reported == 12
-
-    def test_for_agency_not_in_nibrs_month(self, app):
-        q = AgencyParticipation.query
-        q = q.filter(AgencyParticipation.year == 2014)
-        q = q.filter(AgencyParticipation.agency_id == 17427).one()
-        assert q.reported == 1
-        assert q.months_reported == 12
-        assert q.nibrs_reported == 0
-        assert q.nibrs_months_reported == 0
-
-
-class TestParticipationRate:
-    def test_for_state_in_year(self, app):
-        q = ParticipationRate.query
-        q = q.filter(ParticipationRate.year == 2014)
-        q = q.filter(ParticipationRate.state_id == 44).one()
-        assert q.year == 2014
-        assert q.state_id == 44
-        assert q.total_agencies == 62
-        assert q.participating_agencies == 54
-        assert q.participation_rate == pytest.approx(0.870967742)
-        assert q.nibrs_participating_agencies == 52
-        assert q.nibrs_participation_rate == pytest.approx(0.838709677)
-
-    def test_for_county_in_year(self, app):
-        pass
-        q = ParticipationRate.query
-        q = q.filter(ParticipationRate.year == 2014)
-        q = q.filter(ParticipationRate.state_id == 44).one()
-        assert q.year == 2014
-        assert q.state_id == 44
-        assert q.total_agencies == 62
-        assert q.participating_agencies == 54
-        assert q.participation_rate == pytest.approx(0.870967742)
-        assert q.nibrs_participating_agencies == 52
-        assert q.nibrs_participation_rate == pytest.approx(0.838709677)
-
-    def test_for_year(self, app):
-        q = ParticipationRate.query
-        q = q.filter(ParticipationRate.year == 2014)
-        q = q.filter(ParticipationRate.state_id == None)
-        q = q.filter(ParticipationRate.county_id == None).one()
-        assert q.year == 2014
-        assert q.state_id is None
-        assert q.total_agencies == 64
-        assert q.participating_agencies == 54
-        assert q.participation_rate == pytest.approx(0.84375)
-        assert q.nibrs_participating_agencies == 52
-        assert q.nibrs_participation_rate == pytest.approx(0.8125)
-        assert q.total_population == 12142430
-        assert q.participating_population == 1055173
-
-
-class TestCdeAgencies:
-    def test_basic_ref_agency_fields(self, app):
-        a = CdeAgency.query.filter(CdeAgency.agency_id == 22330).one()
-        assert a is not None
-        assert a.ori == 'CA0280400'
-        assert a.legacy_ori == 'CA0280400'
-        assert a.agency_name == 'Yountville'
-        assert a.agency_type_id == 99
-        assert a.agency_type_name == 'Unknown'
-        assert a.state_id == 6
-        assert a.state_abbr == 'CA'
-        assert a.city_id is None
-        assert a.agency_status == 'A'
-        assert a.submitting_agency_id == 23357
-        assert a.submitting_sai == 'CAUCR0001'
-        assert a.submitting_name == 'Department of Justice Criminal Justice Statistics Center'
-        assert a.submitting_state_abbr == 'CA'
-        assert a.dormant_year == 1983
-
-    def test_agency_covering(self, app):
-        a = CdeAgency.query.filter(CdeAgency.agency_id == 17398).one()
-        assert a is not None
-        assert a.current_year == 2014
-        assert a.covered_by_id == 17439
-        assert a.covered_by_ori == 'RIRSP0500'
-        assert a.covered_by_name == 'State Police: Lincoln'
-        assert a.months_reported == 0
-        assert a.nibrs_months_reported == 0
-
-    def test_current_year_and_population(self, app):
-        a = CdeAgency.query.filter(CdeAgency.agency_id == 17385).one()
-        assert a is not None
-        assert a.current_year == 2014
-        assert a.population == 35053
-        assert a.suburban_area_flag == 'Y'
-        assert a.population_group_code == '4'
-        assert a.population_group_desc == 'Cities from 25,000 thru 49,999'
-        assert a.months_reported == 12
-        assert a.nibrs_months_reported == 12
-
-    def test_city_association(self, app):
-        a = CdeAgency.query.filter(CdeAgency.agency_id == 12223).one()
-        assert a is not None
-        assert a.city_id == 6690
-        assert a.city_name == 'Barrington'
-        assert a.state_id == 35
-        assert a.state_abbr == 'NJ'
-
-    def test_staffing_association(self, app):
-        a = CdeAgency.query.filter(CdeAgency.agency_id == 2820).one()
-        assert a is not None
-        assert a.staffing_year == 1990
-        assert a.total_officers == 116
-        assert a.total_civilians == 169
-
-    def test_core_city_flag(self, app):
-        a = CdeAgency.query.filter(CdeAgency.agency_id == 17407).one()
-        assert a is not None
-        assert a.core_city_flag == 'Y'
-
-    def test_primary_county(self, app):
-        a = CdeAgency.query.filter(CdeAgency.agency_id == 17382).one()
-        assert a.primary_county_id == 2402
-        assert a.primary_county == 'Bristol'
-        assert a.primary_county_fips == '44001'
-
-    def test_county_name_append(self, app):
-        a = CdeAgency.query.filter(CdeAgency.ori == 'CA0190000').one()
-        assert a.agency_name == 'Los Angeles County'
-
-    def test_county_not_appended_to_other_agency_type(self, app):
-        a = CdeAgency.query.filter(CdeAgency.ori == 'CA0194200').one()
-        assert a.agency_name == 'Los Angeles'
-
-    def test_revised_rape_start_not_set(self, app):
-        a = CdeAgency.query.filter(CdeAgency.agency_id == 17382).one()
-        assert a is not None
-        assert a.revised_rape_start == 2013
-
-    def test_revised_rape_start_not_set(self, app):
-        a = CdeAgency.query.filter(CdeAgency.agency_id == 17427).one()
-        assert a is not None
-        assert a.revised_rape_start is None
 
     def test_agency_has_unmapped_county(self, app):
         a = CdeAgency.query.filter(CdeAgency.ori == 'RIRSP0000').one()
